Force helm init to install specified version
Impeller will fail if the version of helm on the cluster
is newer then the one specified in the docker image. The user
of impeller should know exactly what version of helm they are
deploying charts with and by adding '--force-upgrade' to helm
the cluster version of helm will match exactly what is in the docker
image.

Docs:
  https://helm.sh/docs/helm/#options-16